feat: moved trigger method body to the baseWrapper
