+++ src/domWrapper.ts
@@ -4,8 +4,7 @@ import BaseWrapper from './baseWrapper'
 import { createWrapperError } from './errorWrapper'
 import WrapperLike from './interfaces/wrapperLike'
 import { DomEventName } from './constants/dom-event-types'
-import { createDOMEvent, TriggerOptions } from './createDomEvent'
-import { nextTick } from 'vue'
+import { TriggerOptions } from './createDomEvent'
 
 export class DOMWrapper<ElementType extends Element>
   extends BaseWrapper<ElementType>
@@ -146,19 +145,6 @@ export class DOMWrapper<ElementType extends Element>
     options?: TriggerOptions
   ): Promise<void>
   async trigger(eventString: string, options?: TriggerOptions) {
-    if (options && options['target']) {
-      throw Error(
-        `[vue-test-utils]: you cannot set the target value of an event. See the notes section ` +
-          `of the docs for more details—` +
-          `https://vue-test-utils.vuejs.org/api/wrapper/trigger.html`
-      )
-    }
-
-    if (this.element && !this.isDisabled()) {
-      const event = createDOMEvent(eventString, options)
-      this.element.dispatchEvent(event)
-    }
-
-    return nextTick()
+    return super.trigger(eventString, options);
   }
 }
